kakao_finance/requirements.txt
yfinance>=0.2.0
pykrx>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
schedule>=1.2.0
